fix: gender neutral title
While I understand that the term "Software Craftsmanship" has been used for years, and conveys a rich meaning, I think that being gender neutral would make our community more inclusive and welcoming for people who are not male. (cf the "Software Crafters" section of https://qz.com/work/1371151/what-happened-to-software-craftsmanship/)

If you agree with this suggestion, I also propose to rename the repository accordingly. FYI: normally, GitHub will automatically redirect visitors from the old URLs to the updated ones, so no link to your repo should be broken.

My intention is not at all to start a debate. Feel free to accept or discard this PR, I will not judge your decision either way, and hope that nobody else will. This is your repository, so you are free to do whatever you want with it. :-)